Crate fireplace_lib

Fireplace Window Manager Library

Fireplace is a modular wayland window manager. This library can be used to build your own window manager by

adding additional functionality to fireplace

replacing functionality of fireplace

or just build upon the concepts of fireplace

As a starting point you might want to fork the fireplace binary to get a functional window manager and use this library to implement your changes. By keeping your changes in the binary or even additional libraries you can make your changes/additions easier accessible to a broder audicence and your fork does not need to maintain a whole window manager with most of the functionality still kept in the unmodified fireplace_lib library.

For a more low level wayland Rust Library take a look at:

wlc (used by fireplace)

or even wayland-server

Concepts

The Callback trait

To understand why fireplace was build the way it was, you need to understand basics of the underlying library wlc:

The Callback Trait

The View Struct

The Output Struct

Output

An Output represents a Display or the X Window when running Fireplace nested in X. In future there may be even more options. An Output is basically everthing wlc and therefor fireplace may render on.

View

A View represents an Application Window in the most basic way. Everything rendered by a wayland client is represented as a View.

Callback

Any object implementing the Callback trait can be used by wlc to drive the compositor. It includes many optionally handled functions that represents events changing the state of the compositor.

Some examples might be:

output_created

output_destroyed

view_created

view_destroyed

pointer_motion

keyboard_key

In wlc you may only use one callback to drive the compositor. Fireplace breaks this restriction by providing Callback implementations that might split the calls onto two or even more other Callbacks and implementations that may wrap existing Callbacks modifing parameters, injecting calls or blocking calls however you want.

Using this functionality every complex process might be split into multiple self-contained so called handlers that can be recombined to create new functionality and can be composed in the end to provide a fully functional window manager.

Handlers

All handlers implement the Callback trait in some way:

directly

by implementing AsWrapper - Wrapping existing handlers

by implementing AsSplit - Splitting calls between up to two handlers

by implementing AsVec - Splitting calls onto an arbitrary number of handlers

All these traits can be found in the callback module

All handlers provided and used by fireplace and a more specific introduction can be found in the handlers module.

Modules

callback
Partial implementations of the Callback trait for easy composability

handlers

utils
Collection of utility functions and macros

Macros

enum_str
Wrapping an enum in enum_str will implement De/Serialize.