Module fireplace_lib::callback

Partial implementations of the Callback trait for easy composability

Using the nightly specialization feature of Rust the structs of this module allow overriding their default implementations of the Callback trait.

Documentation on this subject is very thin, as it is both:

a currently nightly feature with some remaining api/soundness holes

a relatively new nightly feature.

The only official documentation so far is this RFC

Examples on how to use this feature can be found in the struct specific documentation and handler source code.

Structs

Split
Struct wrapping AsSplit Implementations

VecCallback
Struct wrapping AsVecCallback Implementations

Wrapper
Struct wrapping AsWrapper Implementations

Traits

AsSplit
A trait to express that your struct may split Callback calls onto two children.

AsVecCallback
A trait to express that your struct may split Callback calls onto an arbitrary amount of children.

AsWrapper
A trait to express that your struct may forward Callback calls onto a wrapped child.

IntoCallback
A trait to express the ability to consume an object and acquire something that implements Callback