Struct fireplace_lib::callback::Wrapper

pub struct Wrapper<T: AsWrapper>(_);

Struct wrapping AsWrapper Implementations

Because we cannot do specialized implementations for all types implementing AsWrapper, we need to wrap AsWrapper types for the actual implementation.

Take a look at AsWrapper for an example for when to use this.