Trait fireplace_lib::callback::AsSplit

pub trait AsSplit {
    type Callback1: Callback + 'static;
    type Callback2: Callback + 'static;
    fn first_child(&mut self) -> Option<&mut Self::Callback1>;
    fn second_child(&mut self) -> Option<&mut Self::Callback2>;
}

A trait to express that your struct may split Callback calls onto two children.

Knowing that your amount of children will always be two at maximum opens some options for optimizations and makes some edge cases much easier to implement, as it is more common than you might expect. This is main reason it exists next to AsVec. Other than that typing is easier, because both children have different generic types.

You may acquire a Split struct actually implementing Callback via the IntoCallback trait.

Note that fireplace APIs all take IntoCallback types instead of Callback, so you usually don't need to perform the conversation yourself.

Example

use ::wlc::{Callback, View};
use ::fireplace_lib::callback::{IntoCallback, AsSplit, Split};

pub struct ViewSplit<C1: Callback + 'static, C2: Callback + 'static>
{
    pub child1: C1,
    pub child2: C2,
    even: bool,
}

impl<C1, C2> AsSplit for ViewSplit<C1, C2>
    where C1: Callback + 'static, C2: Callback + 'static
{
    type Callback1 = C1;
    type Callback2 = C2;

    fn first_child(&mut self) -> Option<&mut Self::Callback1>
    {
        self.child1
    }
    fn second_child(&mut self) -> Option<&mut Self::Callback2>
    {
        self.child2
    }
}

impl<C1, C2> Callback for Split<ViewSplit<C1, C2>>
    where C1: Callback + 'static, C2: Callback + 'static
{
    fn view_created(&mut self, view: &View) -> bool
    {
        let result = if self.even {
            if let Some(child) = self.first_child() {
                child.view_created(view)
            }
        } else {
            if let Some(child) = self.second_child() {
                child.view_created(view)
            }
        };
        self.even = !self.even;
        result
    }

    // All remaining Callback functions are still send to both children.
// Here specialization kicks in and uses the `default` implementation
of `Split`.
    //
// Note that this example is against the concepts presented in the
handler trait
// All Callbacks should be able to assume all `view_*` methods they may
receive have
    // a corresponding `view_created` even, that has happend earlier.
    // Meaning you never receive events for view's you never saw created.
}

Associated Types

type Callback1: Callback + 'static

Type of the first child

type Callback2: Callback + 'static

Type of the second child

Required Methods

fn first_child(&mut self) -> Option<&mut Self::Callback1>

Returns a mutable reference to your first child, when it exists

fn second_child(&mut self) -> Option<&mut Self::Callback2>

Returns a mutable reference to your second child, when it exists