Trait fireplace_lib::callback::AsVecCallback

pub trait AsVecCallback {
    type Callback: Callback + 'static;
    fn children(&mut self) -> &mut [Self::Callback];
}

A trait to express that your struct may split Callback calls onto an arbitrary amount of children.

You may acquire a VecCallback struct actually implementing Callback via the IntoCallback trait.

Note that fireplace APIs all take IntoCallback types instead of Callback, so you usually don't need to perform the conversation yourself.

A default implementation for std::vec::Vec is provided.

Example

use ::wlc::{Callback, View};
use ::fireplace_lib::callback::{IntoCallback, AsVecCallback, VecCallback};

pub struct CountingViewSplit<C: Callback + 'static>
{
    pub children: Vec<C>,
    pub index: usize,
}

impl<C> AsVecCallback for CountingViewSplit<C>
    where C: Callback + 'static
{
    type Callback = C;

    fn children(&mut self) -> &mut [Self::Callback]
    {
        &mut self.children
    }
}

impl<C> Callback for VecCallback<CountingViewSplit<C>>
    where C: Callback + 'static
{
    fn view_created(&mut self, view: &View) -> bool
    {
        if self.index >= self.children.len() { self.index = 0 }
        self.children()[self.index].view_created(view)
    }

    // All remaining Callback functions are still send to all children.
// Here specialization kicks in and uses the `default` implementation
of `VecCallback`.
    //
// Note that this example is against the concepts presented in the
handler trait
// All Callbacks should be able to assume all `view_*` methods they may
receive have
    // a corresponding `view_created` even, that has happend earlier.
    // Meaning you never receive events for view's you never saw created.
}

Associated Types

type Callback: Callback + 'static

Type of the children

Required Methods

fn children(&mut self) -> &mut [Self::Callback]

Returns a mutable slice of the children