Trait fireplace_lib::callback::AsWrapper

pub trait AsWrapper {
    fn child(&mut self) -> Option<&mut Callback>;
}

A trait to express that your struct may forward Callback calls onto a wrapped child.

You may acquire a Wrapper struct actually implementing Callback via the IntoCallback trait.

Note that fireplace APIs all take IntoCallback types instead of Callback, so you usually don't need to perform the conversation yourself.

Example

use ::wlc::{Callback, View};
use ::fireplace_lib::callback::{IntoCallback, AsWrapper, Wrapper};

pub struct LargeViewFilter<C: Callback + 'static>(pub C);

impl<C> AsWrapper for LargeViewFilter<C>
    where C: Callback + 'static
{
    type Callback = C;

    fn child(&mut self) -> &mut Self::Callback
    {
        &mut self.0
    }
}

impl<C> Callback for Wrapper<LargeViewFilter<C>>
    where C: Callback + 'static
{
    fn view_created(&mut self, view: &View) -> bool
    {
        if view.geometry().size.w * view.geometry().size.h >= 1000000
        {
            self.child().view_created(view)
        }
    }

    // All remaining Callback functions are still send to the child.
// Here specialization kicks in and uses the `default` implementation
of `Wrapper`.
    //
// Note that this example is against the concepts presented in the
handler trait
// All Callbacks should be able to assume all `view_*` methods they may
receive have
    // a corresponding `view_created` even, that has happend earlier.
    // Meaning you never receive events for view's you never saw created.
}

Required Methods

fn child(&mut self) -> Option<&mut Callback>

Returns a mutable reference to the wrapped child, if one exists