Module fireplace_lib::handlers

Handlers

Handlers provide the basic building blocks composed by the compositor to provide it's full functionality.

Dependencies

Handlers might expect other not directly related handlers to be active as well. Because handlers are able to indirectly exchange data through the Store (see below), handlers might expect certain data to be already available from previously run handlers for the same event.

Please take note of the circumstances when designing and combining handlers, or when writing your main module directly in code instead of using a configuration file.

All handlers list their dependencies (if any) in their documentation.

Notable handlers

Store

Handlers may exchange data or track View and Output specific state though the StoreHandler.

When studing the source code of most other handlers understanding how the "Store" may be used is very important.

Workspaces

The WorkspaceHandler is easily the most complex proxy of fireplace, directing all events to related workspaces so they appear to have access to their own Output.

Essentially being an Output multiplexer.

Handling of individual Views is then handled by Modes, like Floating, Fullscreen or BSP (i3-like Binary Split Partitioning).

Render - (might be deactivated by individual features)

The handlers related to the render module enable OpenGL ES drawing and even more high-level drawing engines like conrod directly within the compositor to offer high-performance and/or easy-to-use drawing capabilities.

Reexports

pub use self::keyboard::KeyboardHandler;

pub use self::workspaces::WorkspaceHandler;

pub use self::workspaces::WorkspacesConfig;

Modules

geometry
Handlers related to Output and View geometry

keyboard
Collection of common Keyboard related types.

render
Collection of handlers related to render/draw directly inside of the compositor.

workspaces
Workspace handling

Structs

FocusConfig
Configuration for a FocusHandler

FocusHandler
Handler to direct window focus by given properties

OutputConfig
Output configuration

OutputConfigHandler
Handler to apply output configurations to every newly created Output automatically.

PointerHandler
Handler to move the pointer according to mouse movement

StoreHandler
Handler to initialize a Store for each View and Output

UiConfig
Ui Configuration mostly only available by enabling the ui feature

Traits

Store
Implemented by any View and Output to store arbitrary data.

StoreKey
Implement to store a certain type in the Store.