Module fireplace_lib::handlers::geometry

Handlers related to Output and View geometry

Contains a public interface to restrict and modify the actual used space for View and general and individually. Modes setting their geometry are expected to follow the set values and other handlers may use this interface to change these values.

Structs

GapsConfig
Configuration for the GapsHandler

GapsHandler
Handler to make some gaps around Views

GeometryHandler
Handler that initializes default UsableScreenGeometry and UsableViewGeometry values for created Views and Outputs

InitialViewGeometry
Key for receiving the initial view geometry, that is the size and origin the View has requested at launch, from a Views Sotre.

UsableScreenGeometry
Key for receiving the usable screen geometry, that is the part of the screen not used by UI elements, from an Outputs Store.

UsableViewGeometry
Key for receiving the usable view geometry, that is the part of and area mapped to a View not used by UI elements around it, from a Views Store.

ViewScissor
Value to express that borders of a View shall be shortened