Struct fireplace_lib::handlers::geometry::ViewScissor

pub struct ViewScissor {
    pub up: usize,
    pub right: usize,
    pub down: usize,
    pub left: usize,
}

Value to express that borders of a View shall be shortened

Fields
up: usize

Amount of points the upper border should be moved inside the geometry used

right: usize

Amount of points the right border should be moved inside the geometry used

down: usize

Amount of points the lower border should be moved inside the geometry used

left: usize

Amount of points the left border should be moved inside the geometry used