Module fireplace_lib::handlers::keyboard

Collection of common Keyboard related types.

The KeyboardHandler and KeyHandler may be used for global key handling, while the KeyPattern just describes a common structure of a key action, that may be handled by the compositor. The KeyPattern is therefor often used in other handlers reacting to key presses.

Structs

KeyPattern
Describtion of a key combination that might be handled by the compositor.

KeyboardHandler
Handler to handle global key events

Traits

KeyHandler
Implement to express you may handle a certain key event