Trait fireplace_lib::handlers::keyboard::KeyHandler

pub trait KeyHandler {
    fn handle_key(&mut self, time: u32, view: Option<&View>);
}

Implement to express you may handle a certain key event

Note that this trait makes no assumtions about the actual KeyPattern matched. It only expresses a single action that may may be started.

An Implementation for FnMut is provided

Required Methods

fn handle_key(&mut self, time: u32, view: Option<&View>)

Handle an event

Arguments

time - A value only meaningful in context to other time values, the larger the later an event has happened.

view - The currently focused View if any.