Module fireplace_lib::handlers::render::conrod::provider::statusbar

Types to render a statusbar via conrod on the compositor

Structs

Statusbar
A provider rendering a statusbar

StatusbarConfig
Configuration for a StatusbarHandler describing how the statubar shall look and what it shall display

StatusbarHandler
Handler that initializes a Statusbar provider for every created Output

Time
StatusbarItem for displaying the current time

TimeConfig
Configuration for Time

WorkspaceIndicator
StatusbarItem for displaying the currently active workspace

WorkspaceIndicatorConfig
Configuration for WorkspaceIndicator

Enums

Location

Traits

StatusbarItem
Interface for a statusbar item, that might be rendered on the statusbar