Struct fireplace_lib::handlers::render::conrod::provider::statusbar::WorkspaceIndicatorConfig

pub struct WorkspaceIndicatorConfig {
    pub font: ConfigFont,
    pub alignment: Align,
    pub color: Color,
    pub margin: u32,
}

Configuration for WorkspaceIndicator

Fields
font: ConfigFont

Font to be used for the rendered workspace number and name

alignment: Align

Alignment of the indicator on the statusbar

color: Color

Text color of the rendered workspace number and name

margin: u32

Margin to other StatusbarItems