Trait fireplace_lib::handlers::render::conrod::provider::statusbar::StatusbarItem

pub trait StatusbarItem {
    fn positionable(&mut self,
                    output: &Output,
                    height: f64)
                    -> (Align, Id, Canvas);
    fn render(&mut self, output: &Output, height: f64, ui: &mut UiCell);
}

Interface for a statusbar item, that might be rendered on the statusbar

Required Methods

fn positionable(&mut self, output: &Output, height: f64) -> (Align, Id, Canvas)

Returns a Canvas with an Id, that shall be positioned on the statusbar according to Justify, so the item may be rendered onto of the Canvas without obstructing other StatusbarItems

fn render(&mut self, output: &Output, height: f64, ui: &mut UiCell)

Renders the item using ui with the given height on top of the previously via positionable returned Canvas