Struct fireplace_lib::handlers::render::conrod::ConrodInstance

pub struct ConrodInstance {
    pub ui: Ui,
    pub text_tex: Texture,
    pub text_cache: GlyphCache,
    pub image_map: ImageMap<Texture>,
    // some fields omitted
}

Instance to render ConrodProviders

ConrodInstance manages ConrodProviders as well as the concrete conrod types used for rendering, like Ui, which it also dereferences to for rendering.

Fields
ui: Ui

conrod Ui used for rendering

text_tex: Texture

GL ES Texture used for caching rendered glyphs

text_cache: GlyphCache

Glyph cache for conrod rendering

image_map: ImageMap<Texture>

Map loading images as Textures to be used by conrod via their Ids

Methods

impl ConrodInstance

fn new(dim: Dimensions, logger: Logger) -> ConrodInstance

Initializes a new ConrodInstance

fn register<P: ConrodProvider + 'static>(&mut self, provider: P)

Register a new ConrodProvider to be rendered

fn image_map(&mut self) -> &mut ImageMap<Texture>

Obtain a reference to the ImageMap to store images used by Widgets

fn render(&mut self, output: &mut RenderOutput)

Render on a given RenderOutput