Module fireplace_lib::handlers::render::screenshot

Handler and types related to taking screenshots.

Structs

KeyPatterns
KeyPatterns toggling fullscreen

QueuedScreenshots
Key for receiving the current list of queued screenshot though each Outputs Store.

ScreenshotConfig
Configuration for ScreenshotHandler

ScreenshotHandler
Handler that initializes QueuedScreenshots per Output and automatically calls make_screenshot for every queued screenshot on rendering and automatically saves them as PNG files in the users pictures folder.

Functions

make_screenshot
Create a screenshot of a given Geometry inside a given RenderOutput.