Struct fireplace_lib::handlers::render::screenshot::QueuedScreenshots

pub struct QueuedScreenshots;

Key for receiving the current list of queued screenshot though each Outputs Store.

Meant to be write only. Only append screenshots you want to have taken when the next rendering happens of parts of the specified Output.