Struct fireplace_lib::handlers::render::GLInit

pub struct GLInit<C: Callback + 'static> { /* fields omitted */ }

Handler to load GL functions via egl at runtime for all Output contexts to make GL calls possible.

Please note that you should restore any GL state you may modify to let wlc continue to correctly render any Views.

wlc has currently no way to let use render Views ourselves, so modifing the state directly influences wlcs internal rendering functions, which is likely to render garbage on the screen.

They don't expose any stable API, so even currently working modifications may completely break in the future and should always be avoided.

Take a look at the GraphicsHandler for a stable 2D rendering API

Dependencies

This handler has no dependencies.

Methods

impl<C: Callback + 'static> GLInit<C>

unsafe fn new<I: IntoCallback<C>>(renderer: I) -> GLInit<C>

Create a new GLInit handler. Exposes some unstable APIs.