Struct fireplace_lib::handlers::render::GraphicsRenderer

pub struct GraphicsRenderer<C: Callback> { /* fields omitted */ }

Handler to draw using pistons graphics API for 2D drawing inside the compositor.

It uses and initializes the unstable GLInit handler internally, backing up and restoring the OpenGL state, so you do not need to worry about any of this and may use this stable API for rendering.

The GlGraphics struct is exposed via any Outputs Store to do the rendering.

Dependencies

StoreHandler

Methods

impl<C: Callback + 'static> GraphicsRenderer<C>

fn new<I: IntoCallback<C>>(renderer: I) -> GLInit<Wrapper<GraphicsRenderer<C>>>

Initialize a new GraphicsRenderer