Trait fireplace_lib::handlers::Store

pub trait Store {
    fn insert<T: StoreKey + 'static>(&self, value: T::Value) -> Option<T::Value>;
    fn contains<T: StoreKey + 'static>(&self) -> bool;
    fn get<T: StoreKey + 'static>(&self) -> Option<Rc<RwLock<T::Value>>>;
    fn remove<T: StoreKey + 'static>(&self) -> Option<T::Value>;
}

Implemented by any View and Output to store arbitrary data.

Use structs implementing StoreKey to store any value. Be careful when nesting functions, as get returns a RwLock.

Required Methods

fn insert<T: StoreKey + 'static>(&self, value: T::Value) -> Option<T::Value>

Insert a value into the Store.

Returns any old value that was previously stored under the same key if any.

fn contains<T: StoreKey + 'static>(&self) -> bool

Check if the Store holds any value for the key T.

fn get<T: StoreKey + 'static>(&self) -> Option<Rc<RwLock<T::Value>>>

Get a reference-counted and locked reference to the stored value

Be careful when nesting functions calls or handlers and make sure you don't deadlock.

fn remove<T: StoreKey + 'static>(&self) -> Option<T::Value>

Try to remove and receive the currently stored value

Returns None if any reference or lock is still hold on the value or if no value is stored in the first place.