Trait fireplace_lib::handlers::StoreKey

pub trait StoreKey: Any {
    type Value: Any;
}

Implement to store a certain type in the Store.

Associated Types

type Value: Any

Value is the type to be stored, while the type of the key may be used to add, get and remove the Value type from the Store;