Module fireplace_lib::handlers::workspaces

Workspace handling

Workspace handling is split in three parts in this compositor. At first the WorkspaceHandler dispatches all incoming events to the relevant workspaces (most of the time, only the currently active one), making them able to behave like they manage a completely seperate Output.

A workspace is managed by a Mode, which describes how Views are managed on said workspace.

Examples are Floating or BSP (i3-like Binary Split Partitioning). Modes themselves can be combined by special Modes that dispatch events on certain conditions, examples are Combined, Fullscreen and Switch.

Using just a Mode as an handler without the WorkspaceHandler will effectively result in a functional Output without any workspaces.

Modules

modes
Collection of different Modes

Structs

ActiveWorkspace
Key for receiving the currently active workspace's number and name from any Outputs Store.

KeyPatterns
KeyPatterns triggering workspace related actions

ViewWorkspace
Key for receiving the currently workspace's number and name of any Views Store.

WorkspaceConfig
Configuration for a single Workspace

WorkspaceHandler
Handler that creates and manages Workspaces

WorkspacesConfig
Configuration for the WorkspaceHandler