Module fireplace_lib::handlers::workspaces::modes

Collection of different Modes

Modes handle how Views on an Output or Workspace are layouted and behave. They handle positioning, sizing, reacting to View's requests or may choose to ignore them.

They can be used individually as handlers, but are normally indirectly created and managed by workspaces via their matching configurations.

Reexports

pub use self::bsp::BSP;

pub use self::bsp::BSPConfig;

pub use self::bsp::BSPWrap;

pub use self::bsp::KeyPatterns as BSPKeyPatterns;

Modules

bsp
i3-like Binary Split Partitioning Mode and corresponding handlers and types

Structs

Combined
A Mode that splits View by given filter rules onto two given Modes

CombinedConfig
Configuration for Combined

CombinedKeyPatterns
KeyPatterns toggling modes

Floating
A Mode that does traditional View management.

FloatingConfig
Configuration for Floating

Fullscreen
A Mode that lets you pull one View into fullscreen operation while letting another wrapped Mode handle the rest

FullscreenConfig
Configuration of Fullscreen

FullscreenKeyPatterns
KeyPatterns toggling fullscreen

Switch
A Mode that lets you conviniently switch between different Modes

SwitchConfig
Configuration for Switch

SwitchKeyPatterns
KeyPatterns switching modes

Enums

AnyMode
Enumeration of all possible Modes, makes storing any mode much easier

AnyModeConfig
Configuration for AnyMode

CombinedPredicate
A Predicate may filter given Views

Traits

Mode
Common interface for Modes mostly used by workspaces

Type Definitions

AnyModeWrap
Convinient type alias for a wrapped AnyMode

CombinedWrap
Convinient type alias for a wrapped Combined

FullscreenWrap
Convinient type alias for a wrapped Fullscreen

SwitchWrap
Convinient type alias for a wrapped Switch