Module fireplace_lib::handlers::workspaces::modes::bsp

i3-like Binary Split Partitioning Mode and corresponding handlers and types

Structs

BSP
A tiling Mode that organises Views in a binary tree structure and displays them by deviding the available space

BSPConfig
Configuration for BSP

IndicatorConfig
Configuration for IndicatorsHandler

IndicatorsHandler
A handler used to display border for indication of focus around Views

KeyPatterns
KeyPatterns reorganising the internal tree

Ui
BSP supporting Ui configuration

Enums

Orientation

Type Definitions

BSPWrap
Convinient type alias for a wrapped BSP