Struct fireplace_lib::handlers::workspaces::modes::bsp::BSP

pub struct BSP { /* fields omitted */ }

A tiling Mode that organises Views in a binary tree structure and displays them by deviding the available space