Struct fireplace_lib::handlers::workspaces::modes::bsp::KeyPatterns

pub struct KeyPatterns {
    pub focus_left: Option<KeyPattern>,
    pub focus_right: Option<KeyPattern>,
    pub focus_up: Option<KeyPattern>,
    pub focus_down: Option<KeyPattern>,
    pub move_left: Option<KeyPattern>,
    pub move_right: Option<KeyPattern>,
    pub move_up: Option<KeyPattern>,
    pub move_down: Option<KeyPattern>,
    pub resize_left: Option<KeyPattern>,
    pub resize_right: Option<KeyPattern>,
    pub resize_up: Option<KeyPattern>,
    pub resize_down: Option<KeyPattern>,
    pub toggle_orientation: Option<KeyPattern>,
    pub horizontal_orientation: Option<KeyPattern>,
    pub vertical_orientation: Option<KeyPattern>,
}

KeyPatterns reorganising the internal tree

Fields
focus_left: Option<KeyPattern>

Switch focus to the closest container on the left

focus_right: Option<KeyPattern>

Switch focus to the closest container on the right

focus_up: Option<KeyPattern>

Switch focus to the closest container above

focus_down: Option<KeyPattern>

Switch focus to the closest container below

move_left: Option<KeyPattern>

Move the current container one tile left

move_right: Option<KeyPattern>

Move the current container one tile right

move_up: Option<KeyPattern>

Move the current container one tile up

move_down: Option<KeyPattern>

Move the current container one tile down

resize_left: Option<KeyPattern>

Resize the current container by moving it's border leftwards

resize_right: Option<KeyPattern>

Resize the current container by moving it's border rightwards

resize_up: Option<KeyPattern>

Resize the current container by moving it's border upwards

resize_down: Option<KeyPattern>

Resize the current container by moving it's border downwards

toggle_orientation: Option<KeyPattern>

Toggle the orientation used to split the current container on a new View between Horizontal and Vertical

horizontal_orientation: Option<KeyPattern>

Switch the orientation used to split the current container on a new View to Horizontal

vertical_orientation: Option<KeyPattern>

Switch the orientation used to split the current container on a new View to Vertical