Struct fireplace_lib::handlers::workspaces::modes::Combined

pub struct Combined { /* fields omitted */ }

A Mode that splits View by given filter rules onto two given Modes

Both modes are placed above each other and Predicates may be used to filter which Mode gets it.

A View may also be manually reassigned independenty from the Predicates.